requirements.txt
appdirs==1.4.4
asgiref==3.5.0
astroid==2.5.1
certifi==2020.12.5
channels==3.0.5
chardet==4.0.0
colorama==0.4.4
distlib==0.3.1
Django==3.2.3
django-crispy-forms==1.11.2
filelock==3.0.12
googlemaps==4.4.5
gunicorn==20.1.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
Pillow==8.1.2
pylint==2.7.2
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
sqlite-orm==0.0.2b0